requirements.txt
celery==5.3.4
redis==5.0.1
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
flower==2.0.1
beautifulsoup4==4.12.2
lxml==4.9.3
pandas==2.1.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
drissionpage==4.1.0.18
openai
fake_useragent==1.5.1
python-decouple==3.8
loguru==0.7.2